requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
numpy==1.26.4
sounddevice==0.4.6
openai-whisper==20240930
opencv-python==4.9.0.80
mediapipe==0.10.14